requirements.txt
requests==2.10.0
spacy==1.7.3
nltk==3.2.2
tabulate==0.7.7
numpy==1.12.1
scipy==0.19.9
scikit-learn==0.18.1
matplotlib==2.0.0
python-dateutil==2.6.0
